requirements.txt
Django==1.4.2
Pillow==1.7.8
South==0.7.6
argparse==1.2.1
boto==2.6.0
easy-thumbnails==1.0-alpha-18
ipython==0.13.1
psycopg2==2.4.5
seacucumber==1.5
wsgiref==0.1.2
django-recaptcha==0.0.3
